dpaa_eth: add static qualifier
Add static qualifier for several functions.
Change channel parameter type to uint16_t.

Reviewed-on: http://git.am.freescale.net:8181/18843

+++ drivers/net/ethernet/freescale/dpa/offline_port.h
@@ -39,7 +39,7 @@ struct qman_fq;
 struct fq_duple {
 	struct qman_fq *fqs;
 	int fqs_count;
-	int channel_id;
+	uint16_t channel_id;
 	struct list_head fq_list;
 };
